Refactor anonlink client to use latest blocklib
